+++ src/npc/npc_update/npc_block_logic.cpp
@@ -924,9 +924,10 @@ void NPCBlockLogic(int A, double& tempHit, int& tempHitBlock, float& tempSpeedA,
                                                 }
                                             }
 
-                                            if(((NPC[A]->StandsOnPlayer && !NPC[A].Projectile) || (NPC[A]->IsAShell && NPC[A].Location.SpeedX == 0.0)) && Block[B].tempBlockVehiclePlr > 0)
+                                            if(((NPC[A]->StandsOnPlayer && !NPC[A].Projectile) || (NPC[A]->IsAShell && NPC[A].Location.SpeedX == 0)) && Block[B].tempBlockVehiclePlr > 0)
                                             {
-                                                NPC[A].vehicleYOffset = Block[B].tempBlockVehicleYOffset + NPC[A].Location.Height;
+                                                // FIXME: what are the VB6 semantics of this implicit cast?
+                                                NPC[A].vehicleYOffset = Block[B].tempBlockVehicleYOffset + (int)NPC[A].Location.Height;
                                                 NPC[A].vehiclePlr = Block[B].tempBlockVehiclePlr;
                                                 if(NPC[A].vehiclePlr == 0 && Block[B].tempBlockNpcType == NPCID_VEHICLE)
                                                     NPC[A].TimeLeft = 100;
